wip: 2025-09-05

+++ internal/proof/rebuild.go
@@ -154,7 +154,7 @@ func (segmentTree *ProofSegmentTree) UpdateLayerX(idx int, val common.Hash, laye
 
 		updatedIndices := []int{idx}
 		for idx > 0 {
-			parentIdx := segmenttree.GetParent(idx)
+			parentIdx := segmenttree.GetParent(uint64(idx))
 
 			lChild := tree[2*parentIdx+1]
 			rChild := tree[2*parentIdx+2]
@@ -168,9 +168,9 @@ func (segmentTree *ProofSegmentTree) UpdateLayerX(idx int, val common.Hash, laye
 			// 	rChild.Bytes(),
 			// )
 
-			updatedIndices = append(updatedIndices, parentIdx)
+			updatedIndices = append(updatedIndices, int(parentIdx))
 
-			idx = parentIdx
+			idx = int(parentIdx)
 
 		}
